Updated Tooltip class args to inherit from object.
